Updates readme
There is a much better explanation of what's going on and how to use
this readme document. I also changed it to org mode instead of md,
though this caused me to have to include an SRC emacs lisp block with
just a comment so that it can produce README.el when I
org-babel-do-load-files all the org files. I'll need to fix this at some
point.