Don't wait for R to finish before "freeing" emacs
The default behavior is to wait for R to finish, then print output, then
return control of Emacs. This does not work well for running commands
that take a long time (e.g. bayesian stuff). This new behavior will put
my code in an inferior buffer, send it to the R subprocess, then output
after all the code has run.